Adds Pysa models for flask.Markup
Adds Pysa models to detect XSS issues for flask.Markup

Markup has the __html__ interface that many frameworks use. When
unicode, or html representations are passed to the class, Markup
assumes them to be safe, otherwise converted into unicode
representation.

Hence, defines Markup.__init__ as a TaintSink.

Defines escape function in Markup as a sanitizer for UserControlled
Taint Sources.

The unescape function which de-sanitizes the output of escape function is
declared as a UserControlled TaintSource.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>